[#13]refactor: Props 파일 생성
- 범용성 있는 interface, type들의 관리를 용이하게 하기 위해 Props 파일에서 관리
- 포함 Props: SongProps, LiveProps
